+++ decidim-bulletin_board-app/app/javascript/packs/sandbox/tally.js
@@ -1,5 +1,6 @@
 import $ from "jquery";
 import { TallyComponent, IdentificationKeys } from "../decidim-bulletin_board";
+import { TrusteeWrapperAdapter as DummyTrusteeWrapperAdapter } from "voting-scheme-dummy";
 
 $(() => {
   // UI Elements
@@ -37,6 +38,9 @@ $(() => {
       electionUniqueId,
       trusteeUniqueId: trusteeContext.uniqueId,
       trusteeIdentificationKeys,
+      trusteeWrapperAdapter: new DummyTrusteeWrapperAdapter({
+        trusteeId: trusteeContext.uniqueId,
+      }),
     });
 
     const bindComponentEvents = async () => {